requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
opencv-python
joblib
pillow